Improve hash calculation for ZIP files
