+++ package.json
@@ -72,7 +72,8 @@
     "type": "^2.5.0",
     "untildify": "^4.0.0",
     "uuid": "^8.3.2",
-    "yaml-ast-parser": "0.0.43"
+    "yaml-ast-parser": "0.0.43",
+    "yauzl": "^2.10.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^12.1.4",
